refactor(collector): move the self-syncs into their provider folders
Antigravity's whole self-sync — the native-root check, the sync.lock the subprocess can leave behind, and the spawn itself — together with cursor's maybeSyncCursor, sat in collector.js, the file that schedules them rather than the one that owns them. They move to providers/antigravity/selfSync.js and providers/cursor/selfSync.js, the folders each integration got in #622. The moved bodies are byte-identical to their originals apart from indentation and the one binding described below, and collector.js loses 291 lines.

The process-wide throttle stays in collector.js and is injected into both modules rather than constructed by them. That is not stylistic: one allowance rations cursor and antigravity together because the tokscale cache they guard is a single directory on disk, and the cancellation tests drop only collector.js from the require cache to obtain a fresh one. A throttle constructed inside a provider module would survive that and quietly share state across what those tests take to be separate collectors. Antigravity is handed tokscaleCommand for the same reason, which is also why it needs no lazy require back into collector.js of the kind providers/cursor/auth.js uses to break that cycle.

The antigravity module reaches spawn through the child_process module object instead of destructuring it. Those same tests patch child_process.spawn and then re-require only the collector, so a spawn bound at this module's load time outlives the patch and keeps running the real binary; five tests fail on the destructured form. Going through the module object means stubbing the subprocess no longer depends on knowing the require graph.

collector.js no longer re-exports antigravitySyncLockPath, repairAntigravitySyncLock or removeOwnedAntigravitySyncLock. main.js and the two tests that read them require the provider module directly, since a re-export facade would have left the coupling exactly where it was. Three imports fell unused in collector.js once the syncs left and go with them.

Nothing in the Worker closure moved, so the Hub build marker is unchanged and self-hosted hubs are not asked to redeploy.

AGENTS.md and the antigravity and zed provider notes all claimed collector.js was the only runtime that invokes tokscale. That was already untrue — providers/cursor/auth.js and tokscaleUpdater.js both spawn it — and this change adds a third caller, so the three now say collector.js owns the usage scans rather than making an exhaustive claim.
